Raise exception to stop script
Instead of moving the run variable outside of the function, we add a
flag to the neopixel module and raise an exception when it is true,
so we can just set it to true when we want to stop the script. That
makes it easier for people to use the simulation coming from the
original script, as they do not need to remember to remove the run=1
assignment in the middle of the code.